enforcer-source/requirements.txt
slackclient==2.7.2
absl-py
google.cloud.logging
